Add files via upload
The aim of this prediction model is to provide patient risk stratification for identifying patients who are likely to die during their hospital stay. Data were extracted from the Medical Information Mart for Intensive Care (MIMICIII) database. Clinical prediction models use variables selected because they are thought to be associated (either negatively or positively) with the outcome of interest. The outcome variable is death during the hospital stay.